+++ web/modules/custom/qs_subscription/src/Controller/JudgeController.php
@@ -4,13 +4,15 @@ namespace Drupal\qs_subscription\Controller;
 
 use Drupal\Core\Controller\ControllerBase;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\qs_acl\Service\AccessControl;
 use Drupal\qs_subscription\Service\SubscriptionManager;
+use Drupal\qs_acl\Service\PrivilegeManager;
+use Drupal\Core\Mail\MailManagerInterface;
 use Drupal\Core\Access\AccessResult;
 use Drupal\Core\Session\AccountInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Drupal\qs_subscription\Entity\Subscription;
-use Drupal\Core\Mail\MailManagerInterface;
 
 /**
  * JudgeController.
@@ -28,6 +30,13 @@ class JudgeController extends ControllerBase {
    *
    * @var \Drupal\qs_subscription\Service\SubscriptionManager
    */
+  protected $privilegeManager;
+
+  /**
+   * The Subscription Manager.
+   *
+   * @var \Drupal\qs_subscription\Service\SubscriptionManager
+   */
   protected $subscriptionManager;
 
   /**
@@ -38,12 +47,21 @@ class JudgeController extends ControllerBase {
   protected $mail;
 
   /**
+   * The user Storage.
+   *
+   * @var \Drupal\user\UserStorageInterface
+   */
+  protected $userStorage;
+
+  /**
    * {@inheritdoc}
    */
-  public function __construct(AccessControl $acl, SubscriptionManager $subscription_manager, MailManagerInterface $mail) {
+  public function __construct(EntityTypeManagerInterface $entity_type_manager, AccessControl $acl, PrivilegeManager $privilege_manager, SubscriptionManager $subscription_manager, MailManagerInterface $mail) {
     $this->acl                 = $acl;
+    $this->privilegeManager    = $privilege_manager;
     $this->subscriptionManager = $subscription_manager;
     $this->mail                = $mail;
+    $this->userStorage         = $entity_type_manager->getStorage('user');
   }
 
   /**
@@ -53,7 +71,9 @@ class JudgeController extends ControllerBase {
     // Instantiates this form class.
     return new static(
     // Load customs services used in this class.
+    $container->get('entity_type.manager'),
     $container->get('qs_acl.access_control'),
+    $container->get('qs_acl.privilege_manager'),
     $container->get('qs_subscription.subscription_manager'),
     $container->get('plugin.manager.mail')
     );
@@ -105,9 +125,41 @@ class JudgeController extends ControllerBase {
     // Send email to user when event subscription is approved.
     if ($entity && $entity->bundle() == 'event' && $user && $user->entity) {
       $this->mail->mail('qs_subscription', 'subscription_event_waiting_approval_confirm', $user->entity->getEmail(), $user->entity->getPreferredLangcode(), [
-        'account'   => $user->entity,
-        'event' => $entity,
+        'account' => $user->entity,
+        'event'   => $entity,
       ]);
+
+      $activity = $entity->field_activity->entity;
+
+      // Get all organizers of this activities's event.
+      $query_organizers = $this->privilegeManager->queryPrivilege($activity, 'activity_organizers');
+      $rows = $query_organizers->execute()->fetchAll();
+
+      foreach ($rows as $row) {
+        $ids[$row->user] = $row->user;
+      }
+
+      // Get all maintainer of this activities's event.
+      $query_maintainers = $this->privilegeManager->queryPrivilege($activity, 'activity_maintainers');
+      $rows = $query_maintainers->execute()->fetchAll();
+
+      foreach ($rows as $row) {
+        $ids[$row->user] = $row->user;
+      }
+
+      // Load user with activity_organizers or activity_maintainers privilege(s)
+      // & send them mail.
+      $accounts = NULL;
+      if ($ids) {
+        $accounts = $this->userStorage->loadMultiple($ids);
+
+        foreach ($accounts as $account) {
+          $this->mail->mail('qs_subscription', 'subscription_event_waiting_approval_confirm_organizers', $account->getEmail(), $account->getPreferredLangcode(), [
+            'user'  => $user,
+            'event' => $entity,
+          ]);
+        }
+      }
     }
 
     $confirmed = $this->subscriptionManager->confirm($subscription);
@@ -129,6 +181,17 @@ class JudgeController extends ControllerBase {
    *   JSON formatted response. Contains the status & the declined subscription.
    */
   public function decline(Subscription $subscription) {
+    $entity = $subscription->getEntity();
+    $user = $subscription->getOwner();
+
+    // Send email to user when event subscription is approved.
+    if ($entity && $entity->bundle() == 'event' && $user && $user->entity) {
+      $this->mail->mail('qs_subscription', 'subscription_event_waiting_approval_decline', $user->entity->getEmail(), $user->entity->getPreferredLangcode(), [
+        'account' => $user->entity,
+        'event'   => $entity,
+      ]);
+    }
+
     $declined = $this->subscriptionManager->decline($subscription);
     return new JsonResponse([
       'status'       => TRUE,